# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BookReview {
  id: number;
  bookId: string;
  reviewerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  createdAt: string;
}

export interface BookReviewList {
  bookId: string;
  averageRating: number;
  reviewCount: number;
  reviews: BookReview[];
}

export interface ReviewInput {
  /** @maxLength 80 */
  reviewerName?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  comment: string;
}

export interface ErrorResponse {
  error: string;
}