# lib/api-zod/src/generated/types/bookReviewList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookReview } from './bookReview';

export interface BookReviewList {
  bookId: string;
  averageRating: number;
  reviewCount: number;
  reviews: BookReview[];
}