# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  /** @maxLength 80 */
  reviewerName?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  comment: string;
}